sprites/lighting_controller/pyproject.toml
[project]
name = "lighting-controller"
version = "0.1.10"
description = ""
authors = [
    {name = "Ian King",email = "ian.antking@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)",
    "redis (>=6.2.0,<7.0.0)",
]

[tool.poetry]
packages = [{include = "lighting_controller", from = "src"}]

[tool.poetry.scripts]
lighting_controller = "lighting_controller.__main__:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
